fix(25): merge workspace-reorder into page DndContext to unblock call-drop
Phase-25 regression (verified 2026-05-07T08:25Z): dragging a recording from
the home table onto a workspace row no longer triggered the move. Drag
activation worked (DragOverlay showed "Moving call") but no WorkspaceDropZone
ever received isOver=true.

Root cause: a653f753 added a SECOND DndContext inside WorkspaceSidebarPane to
host the new SortableContext for workspace reorder. dnd-kit's useDroppable
binds to the nearest ancestor DndContext, so WorkspaceDropZone droppables
silently moved into the inner context — invisible to the outer page-level
DndContext where recording rows initiate their drag. Two contexts cannot
share droppables.

Fix: one DndContext per page, both behaviors routed by id-pattern.

- New: src/hooks/useWorkspaceReorder.ts — extracts sensors + a guard
  drag-end handler. The handler is a no-op for any drag whose active.id
  isn't a workspace UUID, so it composes safely with recording/folder
  drop logic in the same context.
- WorkspaceSidebarPane: drop the inner DndContext; keep only SortableContext.
  Removes ~50 lines of sensor/collision/handler boilerplate that now lives
  in the hook.
- TranscriptsNew: plug useWorkspaceReorder into the existing page-level
  DndContext. Call its handler first in handleDragEnd; recording branches
  fall through unchanged.
- RoutingRulesPage: wraps the AppShell in its own DndContext (the page has
  no recording drags) so SortableContext still has an ancestor.

Verified live against localhost:
- /        : reorder works AND call-drop highlights workspace zone +
             toast 'Moved 1 recording' fires
- /rules   : reorder works
- typecheck: 0 errors

Closes Phase 25 GAP-01.
